fix(37-03): break circular DI dependency introduced by BoardType nav gating
ActiveGroupContextService (which QuestBoardContext depends on via
IActiveGroupContext for its tenant query filter) had grown a dependency on
IGroupService, whose repository chain depends back on QuestBoardContext. The
factory-based IActiveGroupContext registration in Program.cs hid this from
.NET's build-time cycle detector, so instead of failing fast it recursed
silently at runtime and the app never got past startup (first surfaced via
SeedShopDataAsync resolving IGroupService before Kestrel could bind).

Move the BoardType lookup onto a new, separate IBoardTypeResolver service
that depends on IActiveGroupContext + IGroupService, while
ActiveGroupContextService itself goes back to depending on only
IHttpContextAccessor. _Layout.cshtml/_Layout.Mobile.cshtml now inject
IBoardTypeResolver instead of asking IActiveGroupContext for the board type.
