+++ QuestBoard.Domain/Interfaces/IActiveGroupContext.cs
@@ -1,5 +1,3 @@
-using QuestBoard.Domain.Enums;
-
 namespace QuestBoard.Domain.Interfaces;
 
 /// <summary>
@@ -9,11 +7,4 @@ namespace QuestBoard.Domain.Interfaces;
 public interface IActiveGroupContext
 {
     int? ActiveGroupId { get; }
-
-    /// <summary>
-    /// Returns the active group's BoardType, or null when no group is active
-    /// (or the active group cannot be resolved). Callers must not assume a
-    /// default value for the null case — it is a distinct state from OneShot/Campaign.
-    /// </summary>
-    Task<BoardType?> GetBoardTypeAsync(CancellationToken token = default);
 }
