+++ QuestBoard.IntegrationTests/Helpers/MutableGroupContext.cs
@@ -4,11 +4,12 @@ using QuestBoard.Domain.Interfaces;
 namespace QuestBoard.IntegrationTests.Helpers;
 
 /// <summary>
-/// Settable implementation of IActiveGroupContext for integration tests.
+/// Settable implementation of IActiveGroupContext and IBoardTypeResolver for integration tests.
 /// Defaults to GroupId = 1 (EuphoriaInn seed group). Tests override as needed.
-/// Registered as Singleton in WebApplicationFactoryBase so test code can mutate it directly.
+/// Registered as Singleton (under both interfaces) in WebApplicationFactoryBase so test code
+/// can mutate it directly.
 /// </summary>
-public class MutableGroupContext : IActiveGroupContext
+public class MutableGroupContext : IActiveGroupContext, IBoardTypeResolver
 {
     public int? ActiveGroupId { get; set; } = 1;
