+++ QuestBoard.Service/Services/ActiveGroupContextService.cs
@@ -1,5 +1,4 @@
 using Microsoft.AspNetCore.Http;
-using QuestBoard.Domain.Enums;
 using QuestBoard.Domain.Interfaces;
 using QuestBoard.Service.Constants;
 
@@ -9,7 +8,7 @@ namespace QuestBoard.Service.Services;
 /// Reads ActiveGroupId from ASP.NET Core Session for HTTP requests.
 /// In Hangfire background threads (no HttpContext), returns null or the override set via SetGroupId.
 /// </summary>
-public class ActiveGroupContextService(IHttpContextAccessor httpContextAccessor, IGroupService groupService) : IActiveGroupContext
+public class ActiveGroupContextService(IHttpContextAccessor httpContextAccessor) : IActiveGroupContext
 {
     private int? _overriddenGroupId;
     private bool _groupIdOverridden;
@@ -23,22 +22,6 @@ public class ActiveGroupContextService(IHttpContextAccessor httpContextAccessor,
         _groupIdOverridden
             ? _overriddenGroupId
             : httpContextAccessor.HttpContext?.Session?.GetInt32(SessionKeys.ActiveGroupId);
-
-    /// <summary>
-    /// Resolves the active group's BoardType via a single group lookup.
-    /// Returns null when no group is active or the active group cannot be found —
-    /// callers (e.g. nav visibility gating) must treat null as its own state, not OneShot.
-    /// </summary>
-    public async Task<BoardType?> GetBoardTypeAsync(CancellationToken token = default)
-    {
-        if (ActiveGroupId is not { } groupId)
-        {
-            return null;
-        }
-
-        var group = await groupService.GetByIdAsync(groupId, token);
-        return group?.BoardType;
-    }
 
     /// <summary>
     /// Called by Hangfire jobs to set the group context before any repository call.
